feat(attachment-explorer): mobile preview, full-page route, message link fix
- Mobile: tapping a row swaps to full-screen preview with a back button;
  search input + filters hide so the preview gets the full viewport.
- Full-page /w/:workspaceId/files mirrors the Memory page chrome and
  reuses the same explorer body via a new ExplorerShell component.
- Sidebar Files entry now navigates to /files instead of toggling the
  modal, matching the Memory link pattern.
- Open-in-stream URLs use ?m= so the destination stream highlights the
  source message (the # form did not match the stream's search-param
  reader).
