+++ apps/frontend/src/components/layout/sidebar/sidebar.tsx
@@ -81,6 +81,7 @@ export function Sidebar({ workspaceId }: SidebarProps) {
   const isScheduledPage = splat === "scheduled" || window.location.pathname.includes("/scheduled")
   const isActivityPage = splat === "activity" || window.location.pathname.endsWith("/activity")
   const isMemoryPage = splat === "memory" || location.pathname.endsWith("/memory")
+  const isFilesPage = splat === "files" || location.pathname.endsWith("/files")
 
   // Build set of streams the user is a member of (for filtering public channels)
   const memberStreamIds = useMemo(() => {
@@ -375,6 +376,7 @@ export function Sidebar({ workspaceId }: SidebarProps) {
               scheduledCount={scheduledCount}
               isActivityPage={isActivityPage}
               isMemoryPage={isMemoryPage}
+              isFilesPage={isFilesPage}
               unreadActivityCount={unreadActivityCount}
             />
           </div>
